Use render mode to render single items at the center of their inventories
