+++ Unity Project/Assets/Plugins/Farrokh Games/Inventory/Editor/Tests/TestProvider.cs
@@ -41,7 +41,6 @@ namespace FarrokhGames.Inventory
 
         public bool DropInventoryItem(IInventoryItem item)
         {
-            UnityEngine.Debug.Log("Item dropped");
             return RemoveInventoryItem(item);
         }
