--- Unity Project/Assets/Plugins/Farrokh Games/Inventory/Internal.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 95e43e411e2b6453cb4bb9b74832035e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
