--- Unity Project/Assets/Plugins/Farrokh Games/Inventory/Internal/AbstractInventoryManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 07941e6fb8915484d82ca0b809ebac74
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
